Extract and fix cw721 approve tests
